cast userid to int in getEffectiveUserId() from auth_proxy
userId's should be int instead of string

+++ phprojekt/library/Phprojekt/Auth/Proxy.php
@@ -105,7 +105,7 @@ class Phprojekt_Auth_Proxy
      */
     public static function getEffectiveUserId()
     {
-        return self::getEffectiveUser()->id;
+        return (int) self::getEffectiveUser()->id;
     }
 
     /**
